DM-8542 -fix setState error -move color utility routines to Color.js -change the symbol dialog to be only one as color dialog. -update drawlayer in color dialog
